# (dev/core#635) Implement local array-cache for use with Redis/Memcache