@totten 2019-01-23
Note: I've taken some bits that were originally copied by each driver (for handling "nack"s).  Those have now been moved to a separate PR, #13500. #13496 is rebased and depends on #13500.