@eileenmcnaughton 2019-01-30
Note that I think we *should* follow this up with a change to caching on the Extension System cache!