@totten 2019-01-30
Agree! There's ~4 that I think we should change.